+++ World/TestIsolatedEmbodiedWorld/TestIsolatedEmbodiedWorld.h
@@ -35,12 +35,12 @@ public:
 //		cout << "Resetting world" << endl;
 	};
 	bool endEvaluation(unsigned long ts) override { return ts >= evaluationTime; };
-	void updateExtraneousWorld(int ts, int visualize) override {  *worldState = ts>=silenceTime ? 1 : 0; };
-	void recordRunningScores(int ts, int visualize) override {
+	void updateExtraneousWorld(unsigned long ts, int visualize) override {  *worldState = ts>=silenceTime ? 1 : 0; };
+	void recordRunningScores(unsigned long ts, int visualize) override {
 		curScore += *worldState==0 ? -1*(*motorEffort) : *motorEffort;
 //		cout << "timestep: " << ts << " score: " << curScore << endl;
 	};
-	void recordSampleScores(int tott, int visualize) override { scores.push_back(curScore); };
+	void recordSampleScores(unsigned long tott, int visualize) override { scores.push_back(curScore); };
 
 	void evaluateOrganism(std::shared_ptr<Organism> currentOrganism, int visualize) override;
 };
